+++ src/math_2d/include/rect.hpp
@@ -9,12 +9,10 @@ namespace levi {
 class vertex;
 class vector2d;
 
-// TODO use floats. x and y have to be vector2d, and width and height have to be
-// size
 class rect {
 public:
   rect();
-  rect(int x, int y, int width, int height);
+  rect(float x, float y, float width, float height);
   /**\return return array of vertices of rectangle, begin from left-top corner
    * oposit-clock-wise. [0,0] in top-left corner.
    * \param z_value z koord for vertices, by default = 0*/
@@ -24,10 +22,10 @@ public:
   /**\return true, if pos is in rectangle, or is on left or top side*/
   bool is_intake_pos(const vector2d &pos);
 
-  int x;
-  int y;
-  int width;
-  int height;
+  float x;
+  float y;
+  float width;
+  float height;
 };
 
 bool is_crossing(const rect &lhs, const rect &rhs);
